Modularize Feature Discovery
